BAH-4667 | Refactor. Move ObservationFormsPanel To observations Folder
